Crate kitchensink_runtime

The Substrate runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

assets_api

Runtime API definition for assets. Runtime API definition for assets.

constants

Constant values used within the runtime. A set of constant values used in substrate runtime.

dynamic_params

Dynamic parameters that can be changed at runtime through the pallet_parameters::set_parameter.

genesis_config_presets

Genesis presets used by this runtime. Genesis Presets for the Kitchensink Runtime

impls

Implementations of some helper traits passed into runtime modules as associated types. Some configurable implementations as associated type for the substrate runtime.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AdjustmentVariable

AlarmInterval

AllianceMaxMembers

AllianceMaxProposals

AllianceMotionDuration

AllyDeposit

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

ApprovalsLimit

AssetConversionOrigin

AssetConversionPalletId

AssetConversionTxHelper

AssetDeposit

AssetRateArguments

AssetRewardsBenchmarkHelper

Benchmark Helper

BagThresholds

BalanceToU256

BasicDeposit

BeefySetIdSessionEntries

BlockHashCount

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

BrokerPalletId

Budget

Burn

ByteDeposit

CandidacyBond

ChallengePeriod

ChildBountyValueMinimum

ClaimPeriod

CodeHashLockupDepositPercent

CollectionDeposit

ConfigDepositBase

CooloffPeriod

CoretimeProvider

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CreationHoldReason

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

CurrencyAdapterDeprecated

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DataDepositPerByte

DealWithFees

DefaultDepositLimit

DelegatedStakingPalletId

DepositBase

DepositFactor

DepositPerByte

DepositPerItem

DesiredMembers

DesiredRunnersUp

DynamicParametersManagerOrigin

ElectionBoundsMultiPhase

ElectionBoundsOnChain

ElectionProviderBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionsPhragmenPalletId

EnactmentPeriod

EnterDepositAmount

EnterDuration

EpochDuration

EthExtraImpl

ExistentialDeposit

ExpectedBlockTime

ExtendDepositAmount

ExtendDuration

FastTrackVotingPeriod

Features

FifoQueueLen

FriendDepositFactor

GraceStrikes

HistoryDepth

ImOnlineUnsignedPriority

IndexDeposit

IntakePeriod

IntoAuthor

ItemAttributesApprovalsLimit

ItemDeposit

LaunchPeriod

LeafVersion

LiquidityWithdrawalFee

LotteryPalletId

MaxActiveValidators

MaxAdditionalFields

MaxAllies

MaxAllowedBytes

MaxAllowedStatements

MaxApprovals

MaxAttributesPerCall

MaxAuthorities

MaxBalance

MaxBids

MaxCalls

MaxCandidates

MaxCollectivesProposalWeight

MaxControllersInDeprecationBatch

MaxDeadlineDuration

MaxElectingVotersSolution

MaxFellows

MaxFriends

MaxGenerateRandom

MaxIntakeWeight

MaxKeys

MaxLockDuration

MaxLocks

MaxNominations

MaxNominators

MaxPayouts

MaxPeerInHeartbeats

MaxPointsToBalance

MaxProposals

MaxQueueLen

MaxRegistrars

MaxReserves

MaxSetIdSessionEntries

MaxSubAccounts

MaxTips

MaxVoters

MaxVotesPerVoter

MaximumMultiplier

MaximumReasonLength

MaximumSchedulerWeight

MbmServiceWeight

MessageQueueServiceWeight

Allocate at most 20% of each block for message processing.

MetadataDepositBase

MetadataDepositPerByte

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinAllowedBytes

MinAllowedStatements

MinBid

MinReceipt

MinVestedTransfer

MinerMaxLength

MinerMaxWeight

MinimumCreditPurchase

MinimumDeposit

MinimumMultiplier

MinimumPeriod

MintMinLiquidity

MixnetNumCoverToCurrentBlocks

MixnetNumCoverToPrevBlocks

MixnetNumRegisterEndSlackBlocks

MixnetNumRegisterStartSlackBlocks

MixnetNumRequestsToCurrentBlocks

MixnetRegistrationPriority

MultiPhaseUnsignedPriority

Native

NewAssetName

NewAssetSymbol

NftFractionalizationPalletId

NisBasePeriod

NisPalletId

NominationPoolsPalletId

NposSolution16

OffchainRandomBalancing

A source of random balance for NposSolver, which is meant to be run by the OCW election miner.

OffchainRepeat

OnChainSeqPhragmen

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PalletTreasuryArguments

PeriodSpend

PoolSetupFee

PostUnbondPoolsWindow

PreimageHoldReason

ProposalDepositOffset

ProposalHoldReason

ProxyDepositBase

ProxyDepositFactor

QueueCount

RecoveryDeposit

ReleaseDelay

ReportLongevity

RetirementPeriod

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SafeModeWhitelistedCalls

Calls that can bypass the safe-mode pallet.

SalaryForRank

Schedule

SessionKeys

SessionsPerEra

SetupAsset

SignedDepositByte

SignedDepositIncreaseFactor

SignedFixedDeposit

SignedPhase

SignedRewardBase

SlashDeferDuration

SlashRewardFraction

SocietyPalletId

SocietyVotingPeriod

SovereignAccountOf

SpendPayoutPeriod

SpendPeriod

StakePoolCreationDeposit

StakingBenchmarkingConfig

StakingRewardsPalletId

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

StatementByteCost

StatementCost

StringLimit

SubAccountDeposit

SubmissionDeposit

Target

TargetBlockFullness

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

ThawThrottle

TipCountdown

TipFindersFee

TipReportDepositBase

TracksInfo

TransactionByteFee

TreasuryAccount

TreasuryPalletId

TxPauseWhitelistedCalls

Calls that cannot be paused by the tx-pause pallet.

U256ToBalance

UndecidingTimeout

UnsignedPhase

UnvestedFundsAllowedWithdrawReasons

UsernameDeposit

Version

VoteLockingPeriod

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

StakerStatus

Representation of the status of a staker.

SudoCall

Contains a variant per dispatchable extrinsic that this pallet has.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

RUNTIME_API_VERSIONS

VERSION

Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Alliance

AllianceConfig

AllianceMotion

AllianceMotionConfig

AssetConversion

AssetConversionMigration

AssetConversionTxPayment

AssetRate

AssetRewards

Assets

AssetsConfig

AssetsFreezer

AssetsFreezerInstance

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Beefy

BeefyConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

Broker

BrokerConfig

CheckedExtrinsic

Extrinsic type that has already been checked.

ChildBounties

Contracts

ConvictionVoting

CoreFellowship

Council

CouncilConfig

DelegatedStaking

Democracy

DemocracyConfig

ElectionProviderMultiPhase

Elections

ElectionsConfig

Executive

Executive: handles dispatch to the various modules.

FastUnstake

Glutton

GluttonConfig

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Lottery

MessageQueue

MetaTx

MetaTxExtension

Mixnet

MixnetConfig

Mmr

MmrLeaf

MultiBlockMigrations

Multiplier

Fee multiplier.

Multisig

NativeAndAssets

NativeAndAssetsFreezer

NftFractionalization

Nfts

Nis

NominationPools

NominationPoolsConfig

Offences

PalletExampleMbms

Parameters

PoolAssets

PoolAssetsConfig

Pov

Preimage

Proxy

RandomnessCollectiveFlip

RankedCollective

RankedPolls

Recovery

Referenda

Remark

Revive

ReviveConfig

RootTesting

SafeMode

SafeModeConfig

Salary

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedPayload

The payload being signed in transactions.

SkipFeelessPayment

Society

SocietyConfig

Staking

StakingConfig

StateTrieMigration

Statement

Sudo

SudoConfig

System

SystemConfig

TasksExample

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TransactionPayment

TransactionPaymentConfig

TransactionStorage

TransactionStorageConfig

Treasury

TreasuryConfig

TxExtension

The TransactionExtension to the basic transaction logic.

TxPause

TxPauseConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UncheckedSignaturePayload

Unchecked signature payload type as expected by this runtime.

Uniques

Utility

VerifySignature

Vesting

VestingConfig

VoterList

Whitelist