Crate kitchensink_runtime

The Substrate runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

assets_api

Runtime API definition for assets. Runtime API definition for assets.

constants

Constant values used within the runtime. A set of constant values used in substrate runtime.

impls

Implementations of some helper traits passed into runtime modules as associated types. Some configurable implementations as associated type for the substrate runtime.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AdjustmentVariable

AlarmInterval

AllianceMaxMembers

AllianceMaxProposals

AllianceMotionDuration

AllowMultiAssetPools

AllyDeposit

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

ApprovalsLimit

AssetConversionOrigin

AssetConversionPalletId

AssetDeposit

BagThresholds

BalanceToU256

BasicDeposit

BetterUnsignedThreshold

BlockHashCount

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

BrokerPalletId

Budget

Burn

CandidacyBond

ChallengePeriod

ChildBountyValueMinimum

ClaimPeriod

CodeHashLockupDepositPercent

CollectionDeposit

ConfigDepositBase

CooloffPeriod

CoreCount

CoretimeProvider

CoretimeRevenue

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

CurrencyAdapter

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DataDepositPerByte

DealWithFees

DefaultDepositLimit

DepositBase

DepositFactor

DepositPerByte

DepositPerItem

DesiredMembers

DesiredRunnersUp

ElectionBoundsMultiPhase

ElectionBoundsOnChain

ElectionProviderBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionsPhragmenPalletId

EnactmentPeriod

EnterDepositAmount

EnterDuration

EpochDuration

ExistentialDeposit

ExpectedBlockTime

ExtendDepositAmount

ExtendDuration

FastTrackVotingPeriod

Features

FieldDeposit

FifoQueueLen

FriendDepositFactor

GraceStrikes

HistoryDepth

ImOnlineUnsignedPriority

IndexDeposit

IntakePeriod

IntoAuthor

ItemAttributesApprovalsLimit

ItemDeposit

LaunchPeriod

LiquidityWithdrawalFee

LotteryPalletId

MaxActiveValidators

MaxAdditionalFields

MaxAllies

MaxAllowedBytes

MaxAllowedStatements

MaxApprovals

MaxAttributesPerCall

MaxAuthorities

MaxBalance

MaxBids

MaxCalls

MaxCandidates

MaxCollectivesProposalWeight

MaxDeadlineDuration

MaxElectingVotersSolution

MaxFellows

MaxFriends

MaxGenerateRandom

MaxIntakeWeight

MaxKeys

MaxLockDuration

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPayouts

MaxPeerInHeartbeats

MaxPointsToBalance

MaxProposals

MaxQueueLen

MaxRegistrars

MaxReserves

MaxSetIdSessionEntries

MaxSubAccounts

MaxTips

MaxVoters

MaxVotesPerVoter

MaximumMultiplier

MaximumReasonLength

MaximumSchedulerWeight

MessageQueueServiceWeight

Allocate at most 20% of each block for message processing.

MetadataDepositBase

MetadataDepositPerByte

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinAllowedBytes

MinAllowedStatements

MinBid

MinReceipt

MinVestedTransfer

MinerMaxLength

MinerMaxWeight

MinimumDeposit

MinimumMultiplier

MinimumPeriod

MintMinLiquidity

MultiPhaseUnsignedPriority

NewAssetName

NewAssetSymbol

NftFractionalizationPalletId

NisBasePeriod

NisPalletId

NominationPoolsPalletId

NposSolution16

OffchainRandomBalancing

A source of random balance for NposSolver, which is meant to be run by the OCW election miner.

OffchainRepeat

OffendingValidatorsThreshold

OnChainSeqPhragmen

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PeriodSpend

PoolSetupFee

PostUnbondPoolsWindow

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

QueueCount

RecoveryDeposit

ReleaseDelay

ReportLongevity

RetirementPeriod

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SafeModeWhitelistedCalls

Calls that can bypass the safe-mode pallet.

SalaryForRank

Schedule

SessionKeys

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedPhase

SignedRewardBase

SlashDeferDuration

SocietyPalletId

SocietyVotingPeriod

SpendPeriod

StakingBenchmarkingConfig

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

StatementByteCost

StatementCost

StringLimit

SubAccountDeposit

SubmissionDeposit

Target

TargetBlockFullness

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

ThawThrottle

TipCountdown

TipFindersFee

TipReportDepositBase

TracksInfo

TransactionByteFee

TreasuryAccount

TreasuryPalletId

TxPauseWhitelistedCalls

Calls that cannot be paused by the tx-pause pallet.

U256ToBalance

UndecidingTimeout

UnsignedPhase

UnvestedFundsAllowedWithdrawReasons

Version

VoteLockingPeriod

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

StakerStatus

Representation of the status of a staker.

SudoCall

Contains a variant per dispatchable extrinsic that this pallet has.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

VERSION

Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Alliance

AllianceConfig

AllianceMotion

AllianceMotionConfig

AssetConversion

AssetConversionTxPayment

AssetRate

AssetTxPayment

Assets

AssetsConfig

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

Broker

CheckedExtrinsic

Extrinsic type that has already been checked.

ChildBounties

Contracts

ConvictionVoting

CoreFellowship

Council

CouncilConfig

Democracy

DemocracyConfig

ElectionProviderMultiPhase

Elections

ElectionsConfig

Executive

Executive: handles dispatch to the various modules.

FastUnstake

GenesisConfigDeprecated

Glutton

GluttonConfig

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Lottery

MessageQueue

Mmr

Multiplier

Fee multiplier.

Multisig

NftFractionalization

Nfts

Nis

NominationPools

NominationPoolsConfig

Offences

PoolAssets

PoolAssetsConfig

Pov

Preimage

Proxy

RandomnessCollectiveFlip

RankedCollective

RankedPolls

Recovery

Referenda

Remark

RootTesting

SafeMode

SafeModeConfig

Salary

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Society

SocietyConfig

Staking

StakingConfig

StateTrieMigration

Statement

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TransactionPayment

TransactionPaymentConfig

TransactionStorage

TransactionStorageConfig

Treasury

TreasuryConfig

TxPause

TxPauseConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

Vesting

VestingConfig

VoterList

Whitelist