Crate kitchensink_runtime

The Substrate runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

Constant values used within the runtime. A set of constant values used in substrate runtime.

impls

Implementations of some helper traits passed into runtime modules as associated types. Some configurable implementations as associated type for the substrate runtime.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AdjustmentVariable

AlarmInterval

AllianceMaxMembers

AllianceMaxProposals

AllianceMotionDuration

AllyDeposit

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetDeposit

BagThresholds

BalanceToU256

BasicDeposit

BetterUnsignedThreshold

BlockHashCount

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidacyBond

CandidateDeposit

ChallengePeriod

ChildBountyValueMinimum

CollectionDeposit

ConfigDepositBase

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

CurrencyAdapter

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DataDepositPerByte

DealWithFees

DeletionQueueDepth

DeletionWeightLimit

DepositBase

DepositFactor

DepositPerByte

DepositPerItem

DesiredMembers

DesiredRunnersUp

ElectionProviderBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionsPhragmenPalletId

EnactmentPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FastTrackVotingPeriod

FieldDeposit

FifoQueueLen

FriendDepositFactor

GenesisConfig

HistoryDepth

IgnoredIssuance

ImOnlineUnsignedPriority

IndexDeposit

IntakePeriod

ItemDeposit

KeyLimit

LaunchPeriod

LotteryPalletId

MaxAdditionalFields

MaxAllies

MaxApprovals

MaxAuthorities

MaxCalls

MaxCandidateIntake

MaxCandidates

MaxElectingVoters

MaxFellows

MaxFounders

MaxFriends

MaxGenerateRandom

MaxIntakeBids

MaxKeys

MaxLockDuration

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPointsToBalance

MaxProposals

MaxQueueLen

MaxRegistrars

MaxReserves

MaxStrikes

MaxSubAccounts

MaxValueSize

MaxVoters

MaximumReasonLength

MaximumSchedulerWeight

MetadataDepositBase

MetadataDepositPerByte

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinFreeze

MinVestedTransfer

MinerMaxLength

MinerMaxWeight

MinimumDeposit

MinimumMultiplier

MinimumPeriod

MultiPhaseUnsignedPriority

NoPreimagePostponement

NominationPoolsPalletId

NposSolution16

OffchainRandomBalancing

A source of random balance for NposSolver, which is meant to be run by the OCW election miner.

OffchainRepeat

OffendingValidatorsThreshold

OnChainSeqPhragmen

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

PeriodSpend

PostUnbondPoolsWindow

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

QueueCount

RecoveryDeposit

ReportLongevity

RetirementPeriod

RewardCurve

RotationPeriod

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

Schedule

SessionKeys

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedPhase

SignedRewardBase

SlashDeferDuration

SocietyPalletId

SpendPeriod

StakingBenchmarkingConfig

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

StringLimit

SubAccountDeposit

SubmissionDeposit

TargetBlockFullness

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TracksInfo

TransactionByteFee

TreasuryPalletId

U256ToBalance

UncleGenerations

UndecidingTimeout

UnsignedPhase

ValueLimit

Version

VoteLockingPeriod

VotingBondBase

VotingBondFactor

VotingPeriod

WrongSideDeduction

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeEvent

StakerStatus

Indicates the initial status of the staker.

SudoCall

Contains one variant per dispatchable that can be called by an extrinsic.

SystemCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

VERSION

Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Alliance

AllianceConfig

AllianceMotion

AllianceMotionConfig

AssetTxPayment

Assets

AssetsConfig

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

CheckedExtrinsic

Extrinsic type that has already been checked.

ChildBounties

Contracts

ConvictionVoting

Council

CouncilConfig

Democracy

DemocracyConfig

ElectionProviderMultiPhase

Elections

ElectionsConfig

Executive

Executive: handles dispatch to the various modules.

Gilt

GiltConfig

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Lottery

Mmr

Multiplier

Fee multiplier.

Multisig

NominationPools

NominationPoolsConfig

Offences

Preimage

Proxy

RandomnessCollectiveFlip

RankedCollective

RankedPolls

Recovery

Referenda

Remark

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Society

SocietyConfig

Staking

StakingConfig

StateTrieMigration

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TransactionPayment

TransactionPaymentConfig

TransactionStorage

TransactionStorageConfig

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

Vesting

VestingConfig

VoterBagsList

Whitelist