Type Alias kitchensink_runtime::AllPalletsWithoutSystem

pub type AllPalletsWithoutSystem = (Utility, Babe, Timestamp, Authorship, Indices, Balances, TransactionPayment, AssetConversionTxPayment, ElectionProviderMultiPhase, Staking, Session, Democracy, Council, TechnicalCommittee, Elections, TechnicalMembership, Grandpa, Treasury, AssetRate, Contracts, Sudo, ImOnline, AuthorityDiscovery, Offences, Historical, RandomnessCollectiveFlip, Identity, Society, Recovery, Vesting, Scheduler, Glutton, Preimage, Proxy, Multisig, Bounties, Tips, Assets, PoolAssets, Beefy, Mmr, MmrLeaf, Lottery, Nis, Uniques, Nfts, NftFractionalization, Salary, CoreFellowship, TransactionStorage, VoterList, StateTrieMigration, ChildBounties, Referenda, Remark, RootTesting, ConvictionVoting, Whitelist, AllianceMotion, Alliance, NominationPools, RankedPolls, RankedCollective, AssetConversion, FastUnstake, MessageQueue, Pov, TxPause, SafeMode, Statement, MultiBlockMigrations, Broker, TasksExample, Mixnet, Parameters, SkipFeelessPayment, PalletExampleMbms, AssetConversionMigration, Revive);

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.